% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{animate_treemap_qc_year}
\alias{animate_treemap_qc_year}
\title{Animate treemap for categorical question over years, write to gif}
\usage{
animate_treemap_qc_year(d, q, gif)
}
\arguments{
\item{d}{data frame from \code{\link{tidy_poll()}}}

\item{q}{question with categorical answers}

\item{gif}{path to output gif}
}
\value{
True
}
\description{
Animate treemap for categorical question over years, write to gif
}
