% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_pctbar_qyn_qc}
\alias{plot_pctbar_qyn_qc}
\title{Plot percent bars by category (from other question) for given yes/no/unsure question}
\usage{
plot_pctbar_qyn_qc(d, q, o_q)
}
\arguments{
\item{d}{data frame from \code{\link{tidy_poll()}}}

\item{q}{question with yes/no/unsure answers}

\item{o_q}{other question with categorical answers}
}
\value{
a ggplot2 graph object
}
\description{
Plot percent bars by category (from other question) for given yes/no/unsure question
}
