% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_pctbar_qyn_year}
\alias{plot_pctbar_qyn_year}
\title{Plot percent bars by year for given yes/no/unsure question}
\usage{
plot_pctbar_qyn_year(d, q)
}
\arguments{
\item{d}{data frame from \code{\link{tidy_poll()}}}

\item{q}{question with yes/no/unsure answers}
}
\value{
a ggplot2 graph object
}
\description{
Plot percent bars by year for given yes/no/unsure question
}
