% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_treemap_qc}
\alias{plot_treemap_qc}
\title{Plot treemap for categorical question}
\usage{
plot_treemap_qc(d, q)
}
\arguments{
\item{d}{data frame from \code{\link{tidy_poll()}}}

\item{q}{question with categorical answers}
}
\value{
a ggplot2 graph object
}
\description{
Plot treemap for categorical question
}
