% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{tidy_poll}
\alias{tidy_poll}
\title{Tidy poll data from a wide to long format for easy querying}
\usage{
tidy_poll(data_xlsx, headers_xlsx, row_end, cols_chr,
  dir_diagnostic_csvs = NULL)
}
\arguments{
\item{data_xlsx}{input path to original poll data Excel spreadsheet}

\item{headers_xlsx}{input path to cleaned headers Excel spreadsheet, with comments aligned to question and no in between text}

\item{row_end}{last row of poll data in original Excel spreadsheet}

\item{cols_chr}{vector of columns that are character}

\item{dir_diagnostic_csvs}{directory to output diagnostic csvs (questions, answers, todo_data-not-numeric)}
}
\value{
tidy data frame
}
\description{
Tidy poll data from a wide to long format for easy querying
}
